Decompile NO3 EntityRoomTransition2
