+++ config/splat.us.stno3.yaml
@@ -54,6 +54,8 @@ segments:
       - [0x3742C, .rodata, 3E134]
       - [0x37440, rodata]
       - [0x37458, rodata]
+      - [0x374DC, .rodata, 3FF00]
+      - [0x374F4, rodata]
       - [0x375E0, .rodata, 46684] # Prize drop
       - [0x37600, .rodata, 46684] # EntityEquipItemDrop
       - [0x37618, rodata]
@@ -64,6 +66,7 @@ segments:
       - [0x377D4, c]
       - [0x3C4EC, c]
       - [0x3E134, c]
+      - [0x3FF00, c]
       - [0x46684, c]
       - [0x56264, c]
       - [0x564B0, c]
