+++ include/game.h
@@ -980,7 +980,7 @@ typedef struct {
     /* 8003C848 */ void (*func_800FE044)(s32, s32);
     /* 8003C84C */ void (*AddToInventory)(u16 itemId, s32 itemCategory);
     /* 8003C850 */ RelicOrb* D_800A8720;
-    /* 8003C854 */ void* func_800FF7B8;
+    /* 8003C854 */ void (* func_800FF7B8)(s32 arg0);
     /* 8003C858 */ s32 (*func_80134714)(s32 arg0, s32 arg1, s32 arg2);
     /* 8003C85C */ s32 (*func_80134678)(s16 arg0, u16 arg1);
     /* 8003C860 */ void (*func_800F53A4)(void);
