+++ src/st/no3/3E134.c
@@ -195,1104 +195,3 @@ INCLUDE_ASM("asm/us/st/no3/nonmatchings/3E134", func_801BEDAC);
 
 // seems to cause a room transition
 INCLUDE_ASM("asm/us/st/no3/nonmatchings/3E134", EntityRoomTransition1);
-
-// seems to cause a room transition
-INCLUDE_ASM("asm/us/st/no3/nonmatchings/3E134", EntityRoomTransition2);
-
-// shows the sword taken from you by Death. Using a different params shows
-// the other items?
-INCLUDE_ASM("asm/us/st/no3/nonmatchings/3E134", EntityDeathStolenItem);
-
-INCLUDE_ASM("asm/us/st/no3/nonmatchings/3E134", EntityUnkId5A);
-
-void EntityUnkId5B(Entity* entity) {
-    Entity* newEntity;
-
-    entity->posX.i.hi = entity[-1].posX.i.hi;
-    entity->posY.i.hi = entity[-1].posY.i.hi;
-
-    switch (entity->step) {
-    case 0:
-        InitializeEntity(D_80180B00);
-        entity->animSet = ANIMSET_OVL(8);
-        entity->palette = 0x2D6;
-        entity->animCurFrame = 0;
-        entity->unk5A = 0x44;
-        break;
-
-    case 1:
-        if (entity->ext.generic.unk7C.u != 0) {
-            switch (entity->ext.generic.unk7C.u) {
-            case 1:
-                AnimateEntity(D_80181B40, entity);
-                break;
-            case 2:
-                AnimateEntity(D_80181B4C, entity);
-                break;
-            case 3:
-                AnimateEntity(D_80181B4C, entity);
-                newEntity = AllocEntity(D_8007D858, &D_8007D858[32]);
-                if (newEntity == NULL) {
-                    break;
-                }
-                CreateEntityFromCurrentEntity(0x5E, newEntity);
-                newEntity->animCurFrame = entity->animCurFrame;
-                newEntity->params = 1;
-                break;
-            }
-        } else {
-            entity->animCurFrame = 0;
-        }
-    }
-    entity->ext.generic.unk7C.s = 0;
-}
-
-void EntityUnkId5E(Entity* entity) {
-    s16 animCurFrame;
-
-    switch (entity->step) {
-    case 0:
-        animCurFrame = entity->animCurFrame;
-        InitializeEntity(D_80180B00);
-        entity->animCurFrame = animCurFrame;
-        entity->animSet = ANIMSET_OVL(8);
-        entity->palette = 0x2D6;
-        entity->unk5A = 0x44;
-        if (entity->params != 0) {
-            entity->unk19 = 8;
-            entity->ext.generic.unk84.U16.unk0 = 0x40;
-        } else {
-            entity->unk19 = 0xC;
-            entity->ext.generic.unk84.U16.unk0 = 0x20;
-        }
-        entity->unk6C = 0x40;
-        entity->blendMode = 0x30;
-        break;
-
-    case 1:
-        if (!(--entity->ext.generic.unk84.U16.unk0)) {
-            DestroyEntity(entity);
-            break;
-        }
-        if (entity->params != 0) {
-            entity->unk6C = (s8)entity->unk6C - 1;
-        } else {
-            entity->unk6C += -2;
-        }
-        break;
-    }
-}
-
-void func_801C13F8() {
-    Entity* entity;
-    s16 temp_s3;
-    s32 i;
-    s8 temp_s4;
-
-    temp_s4 = Random() & 3;
-    temp_s3 = ((Random() & 0xF) << 8) - 0x800;
-    for (i = 0; i < 6; i++) {
-        entity = AllocEntity(D_8007D858, &D_8007D858[MaxEntityCount]);
-        if (entity != NULL) {
-            CreateEntityFromEntity(0x62, g_CurrentEntity, entity);
-            entity->params = 2;
-            entity->ext.generic.unk88.S8.unk1 = 6 - i;
-            entity->ext.generic.unk84.S16.unk0 = temp_s3;
-            entity->ext.generic.unk88.S8.unk0 = temp_s4;
-        }
-    }
-}
-
-// A single "puff" of the warg explosion animation, opaque
-INCLUDE_ASM("asm/us/st/no3/nonmatchings/3E134", EntityWargExplosionPuffOpaque);
-
-s32 Random(void) {
-    g_randomNext = (g_randomNext * 0x01010101) + 1;
-    return g_randomNext >> 0x18;
-}
-
-void Update(void) {
-    s16 i;
-    Entity* entity;
-    s32* unk;
-
-    for (i = 0; i < 0x20; i++) {
-        if (D_801D7DD8[i]) {
-            D_801D7DD8[i]--;
-        }
-    }
-
-    unk = &D_80097410;
-    if (*unk) {
-        if (!--*unk) {
-            g_api.FreePrimitives(D_80097414);
-        }
-    }
-
-    for (entity = &g_Entities[STAGE_ENTITY_START];
-         entity < &g_Entities[TOTAL_ENTITY_COUNT]; entity++) {
-        if (!entity->pfnUpdate)
-            continue;
-
-        if (entity->step) {
-            s32 unk34 = entity->flags;
-            if (unk34 & FLAG_DESTROY_IF_OUT_OF_CAMERA) {
-                s16 posX = i = entity->posX.i.hi;
-                s16 posY = entity->posY.i.hi;
-                if (unk34 & FLAG_DESTROY_IF_BARELY_OUT_OF_CAMERA) {
-                    if ((u16)(posX + 64) > 384 || (u16)(posY + 64) > 352) {
-                        DestroyEntity(entity);
-                        continue;
-                    }
-                } else {
-                    if ((u16)(posX + 128) > 512 || (u16)(posY + 128) > 480) {
-                        DestroyEntity(entity);
-                        continue;
-                    }
-                }
-            }
-
-            if ((unk34 & 0x02000000)) {
-                s16 posY = entity->posY.i.hi + g_Camera.posY.i.hi;
-                s16 test = (LOHU(g_CurrentRoom.vSize) * 256) + 128;
-                if (posY > test) {
-                    DestroyEntity(entity);
-                    continue;
-                }
-            }
-
-            if (unk34 & 0xF) {
-                entity->palette =
-                    D_80181CC0[(entity->nFramesInvincibility << 1) |
-                               (unk34 & 1)];
-                entity->flags--;
-                if ((entity->flags & 0xF) == 0) {
-                    entity->palette = entity->unk6A;
-                    entity->unk6A = 0;
-                }
-            }
-
-            if (!(unk34 & 0x20000000) || (unk34 & 0x10000000) ||
-                ((u16)(entity->posX.i.hi + 64) <= 384) &&
-                    ((u16)(entity->posY.i.hi + 64) <= 352)) {
-                if (!entity->unk58 || (entity->unk58--, unk34 & 0x100000)) {
-                    if (!D_800973FC || unk34 & 0x2100 ||
-                        (unk34 & 0x200 && !(D_8003C8C4 & 3))) {
-                        g_CurrentEntity = entity;
-                        entity->pfnUpdate(entity);
-                        entity->unk44 = 0;
-                        entity->hitFlags = 0;
-                    }
-                }
-            }
-        } else {
-            g_CurrentEntity = entity;
-            entity->pfnUpdate(entity);
-            entity->unk44 = 0;
-            entity->hitFlags = 0;
-        }
-    }
-}
-
-INCLUDE_ASM("asm/us/st/no3/nonmatchings/3E134", func_801C1B78);
-
-INCLUDE_ASM("asm/us/st/no3/nonmatchings/3E134", func_801C1C80);
-
-INCLUDE_ASM("asm/us/st/no3/nonmatchings/3E134", EntityNumericDamage);
-
-void CreateEntityFromLayout(Entity* entity, LayoutEntity* initDesc) {
-    DestroyEntity(entity);
-    entity->entityId = initDesc->entityId & 0x3FF;
-    entity->pfnUpdate = PfnEntityUpdates[entity->entityId - 1];
-    entity->posX.i.hi = initDesc->posX - g_Camera.posX.i.hi;
-    entity->posY.i.hi = initDesc->posY - g_Camera.posY.i.hi;
-    entity->params = initDesc->params;
-    entity->entityRoomIndex = initDesc->entityRoomIndex >> 8;
-    entity->unk68 = (initDesc->entityId >> 0xA) & 7;
-}
-
-void CreateEntityWhenInVerticalRange(LayoutEntity* layoutObj) {
-    s16 yClose;
-    s16 yFar;
-    s16 posY;
-    Entity* entity;
-
-    posY = g_Camera.posY.i.hi;
-    yClose = posY - 0x40;
-    yFar = posY + 0x120;
-    if (yClose < 0) {
-        yClose = 0;
-    }
-
-    posY = layoutObj->posY;
-    if (posY < yClose) {
-        return;
-    }
-
-    if (yFar < posY) {
-        return;
-    }
-
-    switch (layoutObj->entityId & 0xE000) {
-    case 0x0:
-        entity =
-            &g_Entities[STAGE_ENTITY_START + (u8)layoutObj->entityRoomIndex];
-        if (entity->entityId == E_NONE) {
-            CreateEntityFromLayout(entity, layoutObj);
-        }
-        break;
-    case 0x8000:
-        break;
-    case 0xA000:
-        entity =
-            &g_Entities[STAGE_ENTITY_START + (u8)layoutObj->entityRoomIndex];
-        CreateEntityFromLayout(entity, layoutObj);
-        break;
-    }
-}
-
-void CreateEntityWhenInHorizontalRange(LayoutEntity* layoutObj) {
-    s16 xClose;
-    s16 xFar;
-    s16 posX;
-    Entity* entity;
-
-    posX = g_Camera.posX.i.hi;
-    xClose = posX - 0x40;
-    xFar = posX + 0x140;
-    if (xClose < 0) {
-        xClose = 0;
-    }
-
-    posX = layoutObj->posX;
-    if (posX < xClose) {
-        return;
-    }
-
-    if (xFar < posX) {
-        return;
-    }
-
-    switch (layoutObj->entityId & 0xE000) {
-    case 0x0:
-        entity =
-            &g_Entities[STAGE_ENTITY_START + (u8)layoutObj->entityRoomIndex];
-        if (entity->entityId == E_NONE) {
-            CreateEntityFromLayout(entity, layoutObj);
-        }
-        break;
-    case 0x8000:
-        break;
-    case 0xA000:
-        entity =
-            &g_Entities[STAGE_ENTITY_START + (u8)layoutObj->entityRoomIndex];
-        CreateEntityFromLayout(entity, layoutObj);
-        break;
-    }
-}
-
-void func_801C3730(s16 arg0) {
-    while (true) {
-        if ((D_801D7110->posX != (u16)~1) && !(D_801D7110->posX < arg0)) {
-            break;
-        }
-        D_801D7110++;
-    }
-}
-
-void func_801C377C(s16 arg0) {
-    while (true) {
-        if ((D_801D7110->posX == (u16)~0) ||
-            ((arg0 < D_801D7110->posX) && (D_801D7110->posX != (u16)~1))) {
-            D_801D7110--;
-        } else {
-            break;
-        }
-    }
-}
-
-INCLUDE_ASM("asm/us/st/no3/nonmatchings/3E134", func_801C37D4);
-
-INCLUDE_ASM("asm/us/st/no3/nonmatchings/3E134", func_801C38D0);
-
-void func_801C39E4(s16 arg0) {
-    while (true) {
-        if ((D_801D7114[1] != (u16)~1) && !(D_801D7114[1] < arg0)) {
-            break;
-        }
-        D_801D7114 += 5;
-    }
-}
-
-void func_801C3A30(s16 arg0) {
-    while (true) {
-        if (D_801D7114[1] == (u16)~0)
-            D_801D7114 -= 5;
-        else if ((arg0 >= (s32)D_801D7114[1]) || (D_801D7114[1] == (u16)~1))
-            break;
-        else
-            D_801D7114 -= 5;
-    }
-}
-
-INCLUDE_ASM("asm/us/st/no3/nonmatchings/3E134", func_801C3A88);
-
-INCLUDE_ASM("asm/us/st/no3/nonmatchings/3E134", func_801C3B84);
-
-void InitRoomEntities(s32 objLayoutId) {
-    u16* pObjLayoutStart = g_pStObjLayout[objLayoutId];
-    Unkstruct8* currentRoomTileLayout = &g_CurrentRoomTileLayout;
-    s16 temp_s0;
-    s16 arg0;
-    s16 i;
-    u16* temp_v1;
-
-    D_801D7110 = pObjLayoutStart;
-    D_801D7114 = D_80180850[objLayoutId];
-
-    if (*pObjLayoutStart != 0xFFFE) {
-        D_801D7110 = pObjLayoutStart + 1;
-        arg0 = Random() & 0xFF;
-        for (i = 0; true; i++) {
-            temp_v1 = D_801D7110;
-            D_801D7110 = temp_v1 + 1;
-            arg0 -= temp_v1[0];
-            if (arg0 < 0) {
-                break;
-            }
-            D_801D7110 = temp_v1 + 3;
-        }
-        D_801D7110 = (temp_v1[2] << 0x10) + temp_v1[1];
-        D_801D7114 += i * 2 + 2;
-        D_801D7114 = (D_801D7114[1] << 0x10) + D_801D7114[0];
-    }
-    arg0 = currentRoomTileLayout->unkA;
-    temp_s0 = arg0 + 0x140;
-    i = arg0 - 0x40;
-    if (i < 0) {
-        i = 0;
-    }
-
-    D_801D7118 = 0;
-    D_801D711C = 0;
-    func_801C3730(i);
-    func_801C37D4(temp_s0);
-    func_801C39E4(currentRoomTileLayout->unkE + 0x120);
-}
-
-void func_801C37D4(s16);
-void func_801C38D0(s16);
-void func_801C3A88(s16);
-void func_801C3B84(s16);
-
-void func_801C3E10(void) {
-    Unkstruct8* currentRoomTileLayout = &g_CurrentRoomTileLayout;
-
-    if (D_80097908 != 0) {
-        s16 tmp = g_Camera.posX.i.hi;
-        if (D_80097908 > 0)
-            func_801C37D4(tmp + 0x140);
-        else
-            func_801C38D0(tmp - 0x40);
-    }
-
-    if (D_8009790C != 0) {
-        s16 tmp = currentRoomTileLayout->unkE;
-        if (D_8009790C > 0)
-            func_801C3A88(currentRoomTileLayout->unkE + 0x120);
-        else
-            func_801C3B84(tmp - 0x40);
-    }
-}
-
-void CreateEntityFromCurrentEntity(u16 entityId, Entity* entity) {
-    DestroyEntity(entity);
-    entity->entityId = entityId;
-    entity->pfnUpdate = PfnEntityUpdates[entityId - 1];
-    entity->posX.i.hi = g_CurrentEntity->posX.i.hi;
-    entity->posY.i.hi = g_CurrentEntity->posY.i.hi;
-}
-
-void CreateEntityFromEntity(u16 entityId, Entity* source, Entity* entity) {
-    DestroyEntity(entity);
-    entity->entityId = entityId;
-    entity->pfnUpdate = PfnEntityUpdates[entityId - 1];
-    entity->posX.i.hi = source->posX.i.hi;
-    entity->posY.i.hi = source->posY.i.hi;
-}
-
-s32 func_801C3FB4(Entity* e) {
-    s16 diff;
-
-    diff = PLAYER.posX.i.hi - e->posX.i.hi;
-    diff = ABS(diff);
-
-    if (diff >= 17) {
-        diff = 0;
-    } else {
-        diff = PLAYER.posY.i.hi - e->posY.i.hi;
-        diff = ABS(diff);
-        diff = diff < 33;
-    }
-
-    return diff;
-}
-
-INCLUDE_ASM("asm/us/st/no3/nonmatchings/3E134", EntityRedDoor);
-
-void DestroyEntity(Entity* item) {
-    s32 i;
-    s32 length;
-    u32* ptr;
-
-    if (item->flags & FLAG_HAS_PRIMS) {
-        g_api.FreePrimitives(item->primIndex);
-    }
-
-    ptr = (u32*)item;
-    length = sizeof(Entity) / sizeof(s32);
-    for (i = 0; i < length; i++)
-        *ptr++ = 0;
-}
-
-void DestroyEntityFromIndex(s16 index) {
-    Entity* entity = &g_Entities[index];
-
-    while (entity < &D_8007EF1C) {
-        DestroyEntity(entity);
-        entity++;
-    }
-}
-
-void PreventEntityFromRespawning(Entity* entity) {
-    if (entity->entityRoomIndex) {
-        u32 value = (entity->entityRoomIndex - 1);
-        u16 index = value / 32;
-        u16 bit = value % 32;
-        g_entityDestroyed[index] |= 1 << bit;
-    }
-}
-
-#include "st/AnimateEntity.h"
-
-u8 func_801C4E4C(u8 frames[], Entity* self, u8 arg2) {
-    u16 animFrameStart = self->animFrameIdx * 2;
-    u8* var_s1 = &frames[animFrameStart];
-    s16 var_a1 = 0;
-
-    if (self->animFrameDuration == 0) {
-        if (*var_s1 != 0) {
-            if (*var_s1 == 0xFF) {
-                return 0;
-            }
-            self->animFrameDuration = *var_s1++ + (u8)self->ext.stub[0x3F];
-            self->animCurFrame = *var_s1++;
-            self->animFrameIdx++;
-            var_a1 = 128;
-        } else {
-            var_s1 = frames;
-            self->animFrameIdx = 0;
-            self->animFrameDuration = 0;
-            self->ext.stub[0x3F] = (arg2 * Random()) >> 8;
-            self->animFrameDuration = *var_s1++ + (u8)self->ext.stub[0x3F];
-            self->animCurFrame = *var_s1;
-            self->animFrameIdx++;
-            return 0;
-        }
-    }
-    self->animFrameDuration--;
-    self->animCurFrame = var_s1[-1];
-    var_a1 |= 1;
-    return var_a1;
-}
-
-// Absolute distance from g_CurrentEntity to the player in the X Axis
-s32 GetDistanceToPlayerX(void) {
-    s16 xDistance = g_CurrentEntity->posX.i.hi - PLAYER.posX.i.hi;
-
-    if (xDistance < 0) {
-        xDistance = -xDistance;
-    }
-    return xDistance;
-}
-
-// Absolute distance from g_CurrentEntity to the player in the Y Axis
-s32 GetDistanceToPlayerY(void) {
-    s32 yDistance = g_CurrentEntity->posY.i.hi - PLAYER.posY.i.hi;
-
-    if (yDistance < 0) {
-        yDistance = -yDistance;
-    }
-    return yDistance;
-}
-
-s32 GetSideToPlayer(void) {
-    s16 var_a0 = g_CurrentEntity->posX.i.hi > PLAYER.posX.i.hi;
-
-    if (g_CurrentEntity->posY.i.hi > PLAYER.posY.i.hi) {
-        var_a0 |= 2;
-    }
-    return var_a0;
-}
-
-void MoveEntity(void) {
-    g_CurrentEntity->posX.val += g_CurrentEntity->velocityX;
-    g_CurrentEntity->posY.val += g_CurrentEntity->velocityY;
-}
-
-void FallEntity(void) {
-    if (g_CurrentEntity->velocityY < FALL_TERMINAL_VELOCITY) {
-        g_CurrentEntity->velocityY += FALL_GRAVITY;
-    }
-}
-
-s32 func_801C5074(u16* sensors) {
-    Collider col;
-    Collider colBack;
-    s16 x;
-    s16 y;
-    s16 i;
-
-    MoveEntity();
-    FallEntity();
-    if (g_CurrentEntity->velocityY >= 0) {
-        x = g_CurrentEntity->posX.i.hi;
-        y = g_CurrentEntity->posY.i.hi;
-        for (i = 0; i < 4; i++) {
-            x += *sensors++;
-            y += *sensors++;
-            g_api.CheckCollision(x, y, &col, 0);
-            if (col.effects & EFFECT_UNK_8000) {
-                if (i == 1) {
-                    if (col.effects & EFFECT_SOLID) {
-                        g_api.CheckCollision(x, y - 8, &colBack, 0);
-                        if (!(colBack.effects & EFFECT_SOLID)) {
-                            g_CurrentEntity->posY.i.hi =
-                                (u16)g_CurrentEntity->posY.i.hi + 4 + col.unk18;
-                            g_CurrentEntity->velocityX = 0;
-                            g_CurrentEntity->velocityY = 0;
-                            g_CurrentEntity->flags &= ~FLAG_UNK_10000000;
-                            return 1;
-                        }
-                    }
-                    continue;
-                }
-            }
-            if (col.effects & EFFECT_NOTHROUGH && i != 1) {
-                if (col.effects & EFFECT_QUICKSAND) {
-                    g_CurrentEntity->flags &= ~FLAG_UNK_10000000;
-                    return 4;
-                }
-                g_api.CheckCollision(x, y - 8, &colBack, 0);
-                if (!(colBack.effects & EFFECT_SOLID)) {
-                    g_CurrentEntity->posY.i.hi =
-                        g_CurrentEntity->posY.i.hi + col.unk18;
-                    g_CurrentEntity->velocityX = 0;
-                    g_CurrentEntity->velocityY = 0;
-                    g_CurrentEntity->flags &= ~FLAG_UNK_10000000;
-                    return 1;
-                }
-            }
-        }
-    }
-    g_CurrentEntity->flags |= FLAG_UNK_10000000;
-    return 0;
-}
-
-s32 func_801C52EC(s16* posX) {
-    Collider collider;
-    s16 temp2;
-    s16 temp4;
-    s16 x, y;
-
-    g_CurrentEntity->posX.val += g_CurrentEntity->velocityX;
-    temp2 = g_CurrentEntity->posY.i.hi + 3;
-    g_CurrentEntity->posY.i.hi = temp2;
-    x = g_CurrentEntity->posX.i.hi + *posX;
-    posX++;
-    y = temp2 + *posX;
-    g_api.CheckCollision(x, y, &collider, 0);
-    if (!(collider.effects & EFFECT_SOLID)) {
-        return 0;
-    }
-    posX++;
-
-    g_CurrentEntity->posY.i.hi = g_CurrentEntity->posY.i.hi + collider.unk18;
-    if (g_CurrentEntity->velocityX != 0) {
-        if (g_CurrentEntity->velocityX < 0) {
-            temp4 = x - *posX;
-            posX++;
-        } else {
-            temp4 = x + *posX;
-            posX++;
-        }
-        y = y + *posX - 7;
-        g_api.CheckCollision(temp4, y, &collider, 0);
-        if (collider.effects & EFFECT_SOLID) {
-            if ((collider.effects & (EFFECT_UNK_8000 | EFFECT_UNK_0002)) ==
-                EFFECT_UNK_0002) {
-                g_CurrentEntity->posX.val =
-                    g_CurrentEntity->posX.val - g_CurrentEntity->velocityX;
-                g_CurrentEntity->velocityX = 0;
-                return 0xFF;
-            }
-            return 0x61;
-        }
-        y += 15;
-        g_api.CheckCollision(temp4, y, &collider, 0);
-        if (collider.effects & EFFECT_SOLID) {
-            if (collider.effects & EFFECT_UNK_8000) {
-                return 0x61;
-            }
-            return 1;
-        }
-        g_CurrentEntity->posX.val -= g_CurrentEntity->velocityX;
-        g_CurrentEntity->velocityX = 0;
-
-        return 0x80;
-    }
-    return 1;
-}
-
-Entity* AllocEntity(Entity* start, Entity* end) {
-    Entity* current = start;
-    while (current < end) {
-        if (current->entityId == E_NONE) {
-            DestroyEntity(current);
-            return current;
-        }
-
-        current++;
-    }
-    return NULL;
-}
-
-s32 func_801C5534(u8 arg0, s16 arg1) { return D_801820C4[arg0] * arg1; }
-
-s16 func_801C5560(u8 arg0) { return D_801820C4[arg0]; }
-
-void func_801C557C(s32 arg0, s16 arg1) {
-    g_CurrentEntity->velocityX = func_801C5534(arg0, arg1);
-    g_CurrentEntity->velocityY = func_801C5534(arg0 - 0x40, arg1);
-}
-
-u8 func_801C55E8(s16 arg0, s16 arg1) {
-    return ((ratan2(arg1, arg0) >> 4) + 0x40);
-}
-
-u8 func_801C5620(Entity* arg0, Entity* arg1) {
-    s16 a = arg1->posX.i.hi - arg0->posX.i.hi;
-    s16 b = arg1->posY.i.hi - arg0->posY.i.hi;
-    return func_801C55E8(a, b);
-}
-
-u8 func_801C5668(s32 arg0, s32 arg1) {
-    s16 a = (arg0 - (u16)g_CurrentEntity->posX.i.hi);
-    s16 b = (arg1 - (u16)g_CurrentEntity->posY.i.hi);
-    return func_801C55E8(a, b);
-}
-
-u8 func_801C56B0(u8 arg0, u8 arg1, u8 arg2) {
-    u8 var_v0;
-    s8 temp_a2 = arg2 - arg1;
-
-    if (temp_a2 < 0) {
-        var_v0 = -temp_a2;
-    } else {
-        var_v0 = temp_a2;
-    }
-
-    if (var_v0 > arg0) {
-        if (temp_a2 < 0) {
-            var_v0 = arg1 - arg0;
-        } else {
-            var_v0 = arg1 + arg0;
-        }
-
-        return var_v0;
-    }
-
-    return arg2;
-}
-
-// Duplicate
-void func_801C5708(u16 slope, s16 speed) {
-    Entity* entity;
-    s32 moveX;
-    s32 moveY;
-
-    moveX = rcos(slope) * speed;
-    entity = g_CurrentEntity;
-
-    if (moveX < 0) {
-        moveX += 15;
-    }
-
-    entity->velocityX = moveX >> 4;
-
-    moveY = rsin(slope) * speed;
-    entity = g_CurrentEntity;
-
-    if (moveY < 0) {
-        moveY += 15;
-    }
-
-    entity->velocityY = moveY >> 4;
-}
-
-u16 func_801C5794(s16 arg0, s16 arg1) { return ratan2(arg1, arg0); }
-
-u16 func_801C57C4(Entity* a, Entity* b) {
-    s32 diffX = b->posX.i.hi - a->posX.i.hi;
-    s32 diffY = b->posY.i.hi - a->posY.i.hi;
-    return ratan2(diffY, diffX);
-}
-
-u16 func_801C57FC(s32 x, s32 y) {
-    s16 diffX = x - (u16)g_CurrentEntity->posX.i.hi;
-    s16 diffY = y - (u16)g_CurrentEntity->posY.i.hi;
-    return ratan2(diffY, diffX);
-}
-
-u16 func_801C5844(u16 arg0, u16 arg1, u16 arg2) {
-    u16 var_v0 = arg1;
-    u16 temp_a2 = arg2 - arg1;
-    u16 var_v0_2;
-
-    if (temp_a2 & 0x800) {
-        var_v0_2 = (0x800 - temp_a2) & 0x7FF;
-    } else {
-        var_v0_2 = temp_a2;
-    }
-
-    if (var_v0_2 > arg0) {
-        if (temp_a2 & 0x800) {
-            var_v0 = arg1 - arg0;
-        } else {
-            var_v0 = arg1 + arg0;
-        }
-
-        return var_v0;
-    }
-
-    return arg2;
-}
-
-void SetStep(u8 step) {
-    g_CurrentEntity->step = step;
-    g_CurrentEntity->step_s = 0;
-    g_CurrentEntity->animFrameIdx = 0;
-    g_CurrentEntity->animFrameDuration = 0;
-}
-
-void func_801C58C4(u8 step_s) {
-    g_CurrentEntity->step_s = step_s;
-    g_CurrentEntity->animFrameIdx = 0;
-    g_CurrentEntity->animFrameDuration = 0;
-}
-
-void func_801C58E0(u16 arg0, u16 arg1) {
-    Entity* entity;
-
-    if (arg1 != 0) {
-        func_801CAD28(arg1);
-    }
-    if (arg0 == 0xFF) {
-        DestroyEntity(g_CurrentEntity);
-        return;
-    }
-
-    entity = g_CurrentEntity;
-    entity->unk19 = 0;
-    entity->entityId = 2;
-    entity->pfnUpdate = EntityExplosion;
-    entity->params = arg0;
-    entity->animCurFrame = 0;
-    g_CurrentEntity->step = 0;
-    g_CurrentEntity->step_s = 0;
-}
-
-void InitializeEntity(u16 arg0[]) {
-    u16 enemyId;
-    EnemyDef* enemyDef;
-
-    g_CurrentEntity->animSet = *arg0++;
-    g_CurrentEntity->animCurFrame = *arg0++;
-    g_CurrentEntity->unk5A = *arg0++;
-    g_CurrentEntity->palette = *arg0++;
-
-    enemyId = *arg0++;
-    g_CurrentEntity->enemyId = enemyId;
-    enemyDef = &g_api.enemyDefs[enemyId];
-    g_CurrentEntity->hitPoints = enemyDef->hitPoints;
-    g_CurrentEntity->attack = enemyDef->attack;
-    g_CurrentEntity->attackElement = enemyDef->attackElement;
-    g_CurrentEntity->hitboxState = enemyDef->unkC;
-    g_CurrentEntity->hitboxWidth = enemyDef->hitboxWidth;
-    g_CurrentEntity->hitboxHeight = enemyDef->hitboxHeight;
-    g_CurrentEntity->flags = enemyDef->unk24;
-    g_CurrentEntity->hitboxOffX = 0;
-    g_CurrentEntity->hitboxOffY = 0;
-    g_CurrentEntity->step_s = 0;
-    g_CurrentEntity->step++;
-    if (g_CurrentEntity->zPriority == 0) {
-        g_CurrentEntity->zPriority = g_zEntityCenter.S16.unk0 - 0xC;
-    }
-}
-
-void EntityDummy(Entity* arg0) {
-    if (arg0->step == 0) {
-        arg0->step++;
-    }
-}
-
-s32 func_801C5A98(u16* hitSensors, s16 sensorCount) {
-    Collider collider;
-    s16 i;
-    s32 velocityX;
-    u16 temp_a1;
-    s16 x;
-    s16 y;
-
-    velocityX = g_CurrentEntity->velocityX;
-    if (velocityX != 0) {
-        x = g_CurrentEntity->posX.i.hi;
-        y = g_CurrentEntity->posY.i.hi;
-        for (i = 0; i < sensorCount; i++) {
-            if (velocityX < 0) {
-                s16 newX = x + *hitSensors++;
-                x = newX;
-            } else {
-                s16 newX = x - *hitSensors++;
-                x = newX;
-            }
-
-            y += *hitSensors++;
-            g_api.CheckCollision(x, y, &collider, 0);
-            if (collider.effects & EFFECT_UNK_0002 &&
-                ((!(collider.effects & EFFECT_UNK_8000)) || (i != 0))) {
-                return 2;
-            }
-        }
-        return 0;
-    }
-}
-
-void func_801C5BC0(u16* hitSensors, s16 sensorCount) {
-    Collider collider;
-    s16 i;
-    s32 velocityX;
-    s16 x;
-    s16 y;
-
-    velocityX = g_CurrentEntity->velocityX;
-    if (velocityX == 0)
-        return;
-    x = g_CurrentEntity->posX.i.hi;
-    y = g_CurrentEntity->posY.i.hi;
-    for (i = 0; i < sensorCount; i++) {
-        if (velocityX < 0) {
-            x = x + *hitSensors++;
-        } else {
-            x = x - *hitSensors++;
-        }
-
-        y += *hitSensors++;
-        g_api.CheckCollision(x, y, &collider, 0);
-        if (collider.effects & EFFECT_UNK_0002 &&
-            (!(collider.effects & EFFECT_UNK_8000) || i != 0)) {
-            if (velocityX < 0) {
-                g_CurrentEntity->posX.i.hi += collider.unk1C;
-            } else {
-                g_CurrentEntity->posX.i.hi += collider.unk14;
-            }
-            return;
-        }
-    }
-}
-
-INCLUDE_ASM("asm/us/st/no3/nonmatchings/3E134", func_801C5D18);
-
-void ReplaceBreakableWithItemDrop(Entity* self) {
-    u16 params;
-
-    PreventEntityFromRespawning(self);
-
-#if STAGE != STAGE_ST0
-    if (!(g_Status.relics[10] & 2)) {
-        DestroyEntity(self);
-        return;
-    }
-#endif
-
-    params = self->params &= 0xFFF;
-
-    if (params < 0x80) {
-        self->entityId = E_PRIZE_DROP;
-        self->pfnUpdate = (PfnEntityUpdate)EntityPrizeDrop;
-        self->animFrameDuration = 0;
-        self->animFrameIdx = 0;
-    } else {
-        params -= 0x80;
-        self->entityId = E_EQUIP_ITEM_DROP;
-        self->pfnUpdate = (PfnEntityUpdate)EntityEquipItemDrop;
-    }
-
-    self->params = params;
-    self->unk6D = 0x10;
-    self->step = 0;
-}
-
-void func_801C6114(void) {
-    s32 temp_v1;
-    Entity* entity;
-
-    entity = g_CurrentEntity;
-    if (entity->velocityY >= 0) {
-        temp_v1 =
-            entity->ext.generic.unk88.S16.unk0 + entity->ext.generic.unk84.unk;
-        entity->ext.generic.unk84.unk = temp_v1;
-        entity->velocityX = temp_v1;
-        if (temp_v1 == 0x10000 || temp_v1 == -0x10000) {
-            entity->ext.generic.unk88.S16.unk0 =
-                -entity->ext.generic.unk88.S16.unk0;
-        }
-        entity = g_CurrentEntity;
-    }
-
-    if (entity->velocityY < 0x00004000) {
-        entity->velocityY += 0x2000;
-    }
-}
-
-void func_801C6198(u16 arg0) {
-    Collider collider;
-
-    if (g_CurrentEntity->velocityX < 0) {
-        g_api.CheckCollision(g_CurrentEntity->posX.i.hi,
-                             g_CurrentEntity->posY.i.hi - 7, &collider, 0);
-        if (collider.effects & EFFECT_NOTHROUGH) {
-            g_CurrentEntity->velocityY = 0;
-        }
-    }
-
-    g_api.CheckCollision(g_CurrentEntity->posX.i.hi,
-                         g_CurrentEntity->posY.i.hi + 7, &collider, 0);
-
-    if (arg0) {
-        if (!(collider.effects & EFFECT_NOTHROUGH)) {
-            MoveEntity();
-            FallEntity();
-            return;
-        }
-
-        g_CurrentEntity->velocityX = 0;
-        g_CurrentEntity->velocityY = 0;
-
-        if (collider.effects & EFFECT_QUICKSAND) {
-            g_CurrentEntity->posY.val += 0x2000;
-            return;
-        }
-
-        g_CurrentEntity->posY.i.hi += collider.unk18;
-        return;
-    }
-
-    if (!(collider.effects & EFFECT_NOTHROUGH)) {
-        MoveEntity();
-        func_801C6114();
-    }
-}
-
-extern s8 D_801824F0[]; // c_HeartPrizes
-void CollectHeart(u16 arg0) {
-    s32* hearts;
-
-    g_api.PlaySfx(NA_SE_PL_COLLECT_HEART);
-    hearts = &g_Status.hearts;
-    *hearts += D_801824F0[arg0];
-
-    if (g_Status.heartsMax < *hearts) {
-        *hearts = g_Status.heartsMax;
-    }
-
-    DestroyEntity(g_CurrentEntity);
-}
-
-void CollectGold(u16 goldSize) {
-    s32 *gold, *unk;
-    u16 goldSizeIndex;
-
-    g_api.PlaySfx(NA_SE_PL_COLLECT_GOLD);
-    gold = &g_Status.gold;
-    goldSizeIndex = goldSize - 2;
-    *gold += c_GoldPrizes[goldSizeIndex];
-    if (*gold > MAX_GOLD) {
-        *gold = MAX_GOLD;
-    }
-
-    unk = &D_80097410;
-    if (*unk) {
-        g_api.FreePrimitives(D_80097414);
-        *unk = 0;
-    }
-
-    func_801D06FC(D_80182438[goldSizeIndex], 1);
-    DestroyEntity(g_CurrentEntity);
-}
-
-void CollectSubweapon(u16 subWeaponIdx) {
-    Entity* player = &PLAYER;
-    u16 subWeapon;
-
-    g_api.PlaySfx(NA_SE_PL_IT_PICKUP);
-    subWeapon = g_Status.subWeapon;
-    g_Status.subWeapon = D_801823F4[subWeaponIdx];
-
-    if (subWeapon == g_Status.subWeapon) {
-        subWeapon = 1;
-        g_CurrentEntity->unk6D = 0x10;
-    } else {
-        subWeapon = D_80182424[subWeapon];
-        g_CurrentEntity->unk6D = 0x60;
-    }
-
-    if (subWeapon != 0) {
-        g_CurrentEntity->params = subWeapon;
-        g_CurrentEntity->posY.i.hi = player->posY.i.hi + 12;
-        SetStep(7);
-        g_CurrentEntity->velocityY = -0x28000;
-        g_CurrentEntity->animCurFrame = 0;
-        g_CurrentEntity->ext.generic.unk88.S16.unk2 = 5;
-        if (player->facing != 1) {
-            g_CurrentEntity->velocityX = -0x20000;
-            return;
-        }
-        g_CurrentEntity->velocityX = 0x20000;
-        return;
-    }
-    DestroyEntity(g_CurrentEntity);
-}
-
-void CollectHeartVessel(void) {
-    if (g_CurrentPlayableCharacter != PLAYER_ALUCARD) {
-        g_api.PlaySfx(NA_SE_PL_COLLECT_HEART);
-        g_Status.hearts += HEART_VESSEL_RICHTER;
-
-        if (g_Status.heartsMax < g_Status.hearts) {
-            g_Status.hearts = g_Status.heartsMax;
-        }
-    } else {
-        g_api.PlaySfx(NA_SE_PL_COLLECT_HEART);
-        g_api.func_800FE044(HEART_VESSEL_INCREASE, 0x4000);
-    }
-    DestroyEntity(g_CurrentEntity);
-}
-
-void CollectLifeVessel(void) {
-    g_api.PlaySfx(NA_SE_PL_COLLECT_HEART);
-    g_api.func_800FE044(5, 0x8000);
-    DestroyEntity(g_CurrentEntity);
-}
-
-void DestroyCurrentEntity(void) { DestroyEntity(g_CurrentEntity); }
